+++ custom_components/solarcharger/config/config_subentry_custom.py
@@ -27,7 +27,7 @@ from ..const import (
     OPTION_CHARGER_NAME,
     OPTION_GLOBAL_DEFAULTS_ID,
     SENSOR,
-    SENSOR_CHARGER_ALLOCATED_POWER,
+    SENSOR_DELTA_ALLOCATED_POWER,
     SUBENTRY_CHARGER_DEVICE_DOMAIN,
     SUBENTRY_CHARGER_DEVICE_ID,
     SUBENTRY_CHARGER_DEVICE_NAME,
@@ -146,11 +146,11 @@ class AddCustomSubEntryFlowHandler(ConfigSubentryFlow):
                         f"Subentry {SUBENTRY_CHARGER_DEVICE_NAME} not defined"
                     )
 
-                global_defaults_allocated_power = compose_entity_id(
-                    SENSOR, OPTION_GLOBAL_DEFAULTS_ID, SENSOR_CHARGER_ALLOCATED_POWER
+                global_defaults_net_power = compose_entity_id(
+                    SENSOR, OPTION_GLOBAL_DEFAULTS_ID, SENSOR_DELTA_ALLOCATED_POWER
                 )
                 global_defaults_device_entry: DeviceEntry | None = (
-                    self.get_device_entry(global_defaults_allocated_power)
+                    self.get_device_entry(global_defaults_net_power)
                 )
                 if not global_defaults_device_entry:
                     raise ValueError(