+++ custom_components/solarcharger/models/model_context_data.py
@@ -33,6 +33,7 @@ class ContextData:
     is_continue_state: bool = False  # Continue current state
     next_step: ChargeStatus = ChargeStatus.CHARGE_END
     is_enough_power: bool | None = None  # None=not enough data points.
+    net_allocated_power: float = 0
     median_net_allocated_power: float = 0
     data_points: int = 0
 
@@ -56,6 +57,7 @@ class ContextData:
             f"max_allocation_count={self.max_allocation_count}, "
             f"is_enough_power={self.is_enough_power}, "
             f"median_net_allocated_power={self.median_net_allocated_power}, "
+            f"net_allocated_power={self.net_allocated_power}, "
             f"data_points={self.data_points}), "
             f"is_connected={self.is_connected}, "
             f"is_below_charge_limit={self.is_below_charge_limit}, "